+++ CMakeLists.txt
@@ -4,4 +4,4 @@ cmake_minimum_required(VERSION 3.16)
 set(CMAKE_CXX_STANDARD 17) #set the CPP version
 
 include($ENV{IDF_PATH}/tools/cmake/project.cmake)
-project(bt_spp_acceptor_demo)
+project(btsat_v144)
