--- main/at_commands.h
@@ -1,14 +1,0 @@
-/**
- * @file at_commands.h
- * @author Italo Soares (italocjs@live.com)
- * @brief 
- * @version 0.1
- * @date 2023-05-15
- * 
- * @copyright Copyright (c) 2023
- * 
- */
-
-#pragma once
-
-
